Runtime Exceptions
Runtime exceptions differ from Checked exceptions in that the former is not expected to be handled, while the latter does. Runtime exceptions are errors that appear during runtime, hence the name, and are exceptions that the program will be unlikely to recover from. Runtime errors can be handled but it is uncommon to do so, as fixing and eliminating runtime exceptions are preferable in nearly all cases.